feat: diversos

+++ src/server/database/migrations/0002_create_municipios.ts
@@ -7,7 +7,7 @@ export async function up(knex: Knex) {
     return knex
         .schema
         .createTable(ETableNames.municipio, table => {
-            table.bigIncrements('id').primary().index()
+            table.integer('id').primary().index()
             table.string('name', 75).checkLength('<=', 75).index().notNullable()
             table.string('state', 2).checkLength('=', 2).index().notNullable()
